+++ site/app/models/gradeable/AutogradingConfig.php
@@ -55,9 +55,7 @@ class AutogradingConfig extends AbstractModel {
     protected $one_part_only;
 
     /** @prop @var array Array of notebook objects */
-    private $notebook = [];
-    /** @prop @var AbstractGradingInput[] Grading input configs for all new types of gradeable input*/
-    private $inputs = [];
+    protected $notebook_config = [];
     /** @prop @var AutogradingTestcase[] Cut-down information about autograding test cases*/
     private $testcases = [];
 
@@ -86,8 +84,6 @@ class AutogradingConfig extends AbstractModel {
     protected $total_hidden_non_extra_credit = 0;
     /** @prop @var int Total number of hidden extra-credit points for all test cases */
     protected $total_hidden_extra_credit = 0;
-    /** @prop @var bool have we parsed the notebook config and created a userspecific notebook */
-    private $is_notebook_parsed = false;
 
 
     public function __construct(Core $core, array $details) {
@@ -114,35 +110,7 @@ class AutogradingConfig extends AbstractModel {
         $this->required_capabilities = $details['required_capabilities'] ?? 'default';
         $this->max_possible_grading_time = $details['max_possible_grading_time'] ?? -1;
 
-    
-        if (isset($details['testcases'])) {
-            foreach ($details['testcases'] as $idx => $testcase_details) {
-                $testcase = new AutogradingTestcase($this->core, $testcase_details, $idx);
-
-                // Accumulate only the positive points
-                $points = $testcase->getPoints();
-                if ($points >= 0.0) {
-                    if ($testcase->isHidden()) {
-                        if ($testcase->isExtraCredit()) {
-                            $this->total_hidden_extra_credit += $points;
-                        }
-                        else {
-                            $this->total_hidden_non_extra_credit += $points;
-                        }
-                    }
-                    else {
-                        if ($testcase->isExtraCredit()) {
-                            $this->total_non_hidden_extra_credit += $points;
-                        }
-                        else {
-                            $this->total_non_hidden_non_extra_credit += $points;
-                        }
-                    }
-                }
-
-                $this->testcases[$idx] = $testcase;
-            }
-        }
+        $this->parseTestCases($details);
 
         if (isset($details['early_submission_incentive'])) {
             $this->early_submission_incentive = true;
@@ -154,13 +122,8 @@ class AutogradingConfig extends AbstractModel {
             }
         }
 
-
         if (isset($details['notebook'])) {
-            $this->notebook = [
-                'notebook' => $details['notebook'],
-                'item_pool' => []
-            ];
-
+            $this->notebook_config = $details['notebook'];
             $this->notebook_gradeable = true;
         }
 
@@ -184,162 +147,10 @@ class AutogradingConfig extends AbstractModel {
     }
 
 
-    //TODO: remove once we can store notebook in its own model
-    public function parseNotebook(array $details): void {
-        // Setup $this->notebook
-        $actual_input = array();
-        $this->notebook = [];
-
-        // For each item in the notebook array inside the $details collect data and assign to variables in
-        // $this->notebook
-        for ($i = 0; $i < count($details['notebook']); $i++) {
-            $notebook_cell = $details['notebook'][$i];
-            $do_add = true;
-
-            // If cell is of markdown type then figure out if it is markdown_string or markdown_file and pass this
-            // markdown forward as 'data' as opposed to 'string' or 'file'
-            if (
-                isset($notebook_cell['type'])
-                && $notebook_cell['type'] === 'markdown'
-            ) {
-                $markdown = $this->getMarkdownData($notebook_cell);
-
-                // Remove string or file from $notebook_cell
-                unset($notebook_cell['markdown_string']);
-                unset($notebook_cell['markdown_file']);
-
-                // Read as data
-                $notebook_cell['markdown_data'] = $markdown;
-
-                // If next entry is an input type, we assign this as a label - otherwise it is plain markdown
-                if ($i < count($details['notebook']) - 1) {
-                    $next_cell = &$details['notebook'][$i + 1];
-                    if (
-                        isset($next_cell['type'])
-                        && ($next_cell['type'] == 'short_answer' || $next_cell['type'] == 'multiple_choice')
-                    ) {
-                        $next_cell['label'] = $markdown;
-                        // Do not add current cell to notebook, since it is embedded in the label
-                        $do_add = false;
-                    }
-                }
-            }
-            elseif (
-                $notebook_cell['type'] === 'short_answer'
-                && !empty($notebook_cell['programming_language'])
-                && empty($notebook_cell['codemirror_mode'])
-            ) {
-                $notebook_cell['codemirror_mode'] = Utils::getCodeMirrorMode($notebook_cell['programming_language']);
-            }
-
-            // Add this cell $this->notebook
-            if ($do_add) {
-                $this->notebook[] = $notebook_cell;
-            }
-
-            // If cell is a type of input add it to the $actual_inputs array
-            if (in_array($notebook_cell['type'], ['short_answer', 'multiple_choice'])) {
-                $actual_input[] = $notebook_cell;
-            }
-        }
-        
-
-        // Setup $this->inputs
-        for ($i = 0; $i < count($actual_input); $i++) {
-            if ($actual_input[$i]['type'] == 'short_answer') {
-                // If programming language is set then this is a codebox, else regular textbox
-                if (isset($actual_input[$i]['programming_language'])) {
-                    $this->inputs[$i] = new SubmissionCodeBox($this->core, $actual_input[$i]);
-                }
-                else {
-                    $this->inputs[$i] = new SubmissionTextBox($this->core, $actual_input[$i]);
-                }
-            }
-            elseif ($actual_input[$i]['type'] == 'multiple_choice') {
-                $actual_input[$i]['allow_multiple'] = $actual_input[$i]['allow_multiple'] ?? false;
-
-                $this->inputs[$i] = new SubmissionMultipleChoice($this->core, $actual_input[$i]);
-            }
-        }
-
-        if (isset($details['item_pool'])) {
-            $this->notebook['item_pool'] = $details['item_pool'];
-        }
-
-        // defaults num of parts to 1 if value is not set
-        $num_parts = count($details['part_names'] ?? [1]);
-        $this->one_part_only = $details['one_part_only'] ?? false;
-
-        // Get all of the part names
-        for ($i = 1; $i <= $num_parts; $i++) {
-            $j = $i - 1;
-            if (
-                isset($details['part_names'])
-                && isset($details['part_names'][$j])
-                && trim($details['part_names'][$j]) !== ""
-            ) {
-                $this->part_names[$i] = $details['part_names'][$j];
-            }
-            else {
-                $this->part_names[$i] = "Part " . $i;
-            }
-        }
-
-
-        if (isset($details['testcases'])) {
-            foreach ($details['testcases'] as $idx => $testcase_details) {
-                $testcase = new AutogradingTestcase($this->core, $testcase_details, $idx);
-
-                // Accumulate only the positive points
-                $points = $testcase->getPoints();
-                if ($points >= 0.0) {
-                    if ($testcase->isHidden()) {
-                        if ($testcase->isExtraCredit()) {
-                            $this->total_hidden_extra_credit += $points;
-                        }
-                        else {
-                            $this->total_hidden_non_extra_credit += $points;
-                        }
-                    }
-                    else {
-                        if ($testcase->isExtraCredit()) {
-                            $this->total_non_hidden_extra_credit += $points;
-                        }
-                        else {
-                            $this->total_non_hidden_non_extra_credit += $points;
-                        }
-                    }
-                }
-
-                $this->testcases[$idx] = $testcase;
-            }
-        }
-    }
-
-
-    private function getMarkdownData($cell) {
-        // If markdown_string is set then just return that
-        if (isset($cell['markdown_string'])) {
-            return $cell['markdown_string'];
-        }
-        elseif (isset($cell['markdown_file'])) {
-            // Else if markdown_file is set then read the file and return its contents
-            // TODO: Implement reading from markdown_file and passing that along
-            throw new NotImplementedException("Reading from a markdown_file is not yet implemented.");
-        }
-        else {
-            // Else something unexpected happened
-            throw new \InvalidArgumentException("An error occured parsing notebook data.\n" .
-                "Markdown configuration may only specify one of 'markdown_string' or 'markdown_file'");
-        }
-    }
-
     public function toArray() {
         $details = parent::toArray();
 
         $details['testcases'] = parent::parseObject($this->testcases);
-        $details['inputs'] = parent::parseObject($this->inputs);
-
         return $details;
     }
 
@@ -351,44 +162,42 @@ class AutogradingConfig extends AbstractModel {
         return $this->testcases;
     }
 
-    /**
-     * Gets the abstract inputs for this configuration
-     * @return AbstractGradeableInput[]
-     */
-    public function getInputs() {
-        return $this->inputs;
+
+    public function getNotebook() {
+        return $this->notebook_config;
     }
 
-    /**
-    * creates the notebook by parsing the notebook config and creating a userspecific notebook
-    * @param string $gradeable_id
-    * @param string $user_id
-    * @return UserspecificNotebook|array returns empty array if this gradeable is not a notebook gradeable
-    */
-    public function getNotebook(string $gradeable_id, string $user_id) {
-        if ($this->is_notebook_parsed || !$this->notebook_gradeable) {
-            return $this->notebook;
+
+    public function parseTestCases(array $details): void {
+        if (!isset($details['testcases'])) {
+            return;
         }
-        $notebook_model = new UserSpecificNotebook(
-            $this->core,
-            $this->notebook,
-            $gradeable_id,
-            $user_id
-        );
+        foreach ($details['testcases'] as $idx => $testcase_details) {
+            $testcase = new AutogradingTestcase($this->core, $testcase_details, $idx);
 
-        $this->parseNotebook([
-            'notebook' => $notebook_model->getNotebookConfig(),
-            'testcases' => $notebook_model->getTestCases()
-        ]);
+            // Accumulate only the positive points
+            $points = $testcase->getPoints();
+            if ($points >= 0.0) {
+                if ($testcase->isHidden()) {
+                    if ($testcase->isExtraCredit()) {
+                        $this->total_hidden_extra_credit += $points;
+                    }
+                    else {
+                        $this->total_hidden_non_extra_credit += $points;
+                    }
+                }
+                else {
+                    if ($testcase->isExtraCredit()) {
+                        $this->total_non_hidden_extra_credit += $points;
+                    }
+                    else {
+                        $this->total_non_hidden_non_extra_credit += $points;
+                    }
+                }
+            }
 
-        $this->is_notebook_parsed = true;
-
-        //now the autogradingconfig has reparsed the notebook,
-        //update the model's representation
-        $notebook_model->setNotebookConfig($this->notebook);
-        $this->notebook = $notebook_model;
-
-        return $notebook_model;
+            $this->testcases[$idx] = $testcase;
+        }
     }
 
     /**
@@ -407,13 +216,6 @@ class AutogradingConfig extends AbstractModel {
         return count($this->getPartNames());
     }
 
-    /**
-     * Gets the number of inputs on the submission page
-     * @return int
-     */
-    public function getNumInputs() {
-        return count($this->getInputs());
-    }
 
     /**
      * Gets the number of non-hidden points possible for this assignment (including extra credit)
