+++ site/app/models/gradeable/GradedGradeable.php
@@ -269,118 +269,6 @@ class GradedGradeable extends AbstractModel {
     }
 
 
-    /**
-     * Gets a new 'notebook' which contains information about most recent submissions
-     *
-     * @return array An updated 'notebook' which has the most recent submission data entered into the
-     * 'recent_submission' key for each input item inside the notebook.  If there haven't been any submissions,
-     * then 'recent_submission' is populated with 'initial_value' if one exists, otherwise it will be
-     * blank.
-     */
-    public function getUpdatedNotebook(array $newNotebook): array {
-        foreach ($newNotebook as $notebookKey => $notebookVal) {
-            if (isset($notebookVal['type'])) {
-                if ($notebookVal['type'] == "short_answer") {
-                    // If no previous submissions set string to default initial_value
-                    if ($this->getAutoGradedGradeable()->getHighestVersion() == 0) {
-                        $recentSubmission = $notebookVal['initial_value'] ?? "";
-                    }
-                    else {
-                        // Else there has been a previous submission try to get it
-                        try {
-                            // Try to get the most recent submission
-                            $recentSubmission = $this->getRecentSubmissionContents($notebookVal['filename']);
-                        }
-                        catch (AuthorizationException $e) {
-                            // If the user lacked permission then just set to default instructor provided string
-                            $recentSubmission = $notebookVal['initial_value'] ?? "";
-                        }
-                    }
-
-                    // Add field to the array
-                    $newNotebook[$notebookKey]['recent_submission'] = $recentSubmission;
-                }
-                elseif ($notebookVal['type'] == "multiple_choice") {
-                    // If no previous submissions do nothing, else there has been, so try and get it
-                    if ($this->getAutoGradedGradeable()->getHighestVersion() == 0) {
-                        continue;
-                    }
-                    else {
-                        try {
-                            // Try to get the most recent submission
-                            $recentSubmission = $this->getRecentSubmissionContents($notebookVal['filename']);
-
-                            // Add field to the array
-                            $newNotebook[$notebookKey]['recent_submission'] = $recentSubmission;
-                        }
-                        catch (AuthorizationException $e) {
-                            // If failed to get the most recent submission then skip
-                            continue;
-                        }
-                    }
-                }
-            }
-        }
-
-        // Operate on notebook to add prev_submission field to inputs
-        return $newNotebook;
-    }
-
-
-    /**
-     * Get the data from the student's most recent submission
-     *
-     * @param $filename Name of the file to collect the data out of
-     * @throws AuthorizationException if the user lacks permissions to read the submissions file
-     * @throws FileNotFoundException if file with passed filename could not be found
-     * @throws IOException if there was an error reading contents from the file
-     * @return string if successful returns the contents of a students most recent submission
-     */
-    private function getRecentSubmissionContents($filename) {
-
-        // Get items in path to student's submission folder
-        $course_path = $this->core->getConfig()->getCoursePath();
-        $gradable_dir = $this->getGradeableId();
-        $student_id = $this->core->getUser()->getId();
-        $version = $this->getAutoGradedGradeable()->getHighestVersion();
-
-        // Join path items
-        $complete_file_path = FileUtils::joinPaths(
-            $course_path,
-            'submissions',
-            $gradable_dir,
-            $student_id,
-            $version,
-            $filename
-        );
-
-        // Check if the user has permission to access this submission
-        $isAuthorized = $this->core->getAccess()->canI('path.read', ["dir" => "submissions", "path" => $complete_file_path]);
-
-        // If user lacks permission to get the submission contents throw Auth exception
-        if (!$isAuthorized) {
-            throw new AuthorizationException("The user lacks permissions to access this data.");
-        }
-
-        // If desired file does not exist in the most recent submission directory throw exception
-        if (!file_exists($complete_file_path)) {
-            throw new FileNotFoundException("Unable to locate submission file.");
-        }
-
-        // Read file contents into string
-        $file_contents = file_get_contents($complete_file_path);
-
-        // If file_contents is False an error has occured
-        if ($file_contents === false) {
-            throw new IOException("An error occurred retrieving submission contents.");
-        }
-
-        // Remove trailing newline
-        $file_contents = rtrim($file_contents, "\n");
-
-        return $file_contents;
-    }
-
     public function hasSubmission() {
         return $this->gradeable->hasSubmission($this->submitter);
     }
