+++ site/app/models/notebook/SubmissionCodeBox.php
@@ -1,10 +1,10 @@
 <?php
 
-namespace app\models\gradeable;
+namespace app\models\notebook;
 
 use app\libraries\Core;
 use app\libraries\Utils;
-use app\models\gradeable\AbstractTextBox;
+use app\models\notebook\AbstractTextBox;
 
 /**
  * Information required to load code box submissions on the submission page
