+++ site/app/models/notebook/SubmissionTextBox.php
@@ -1,13 +1,13 @@
 <?php
 
-namespace app\models\gradeable;
+namespace app\models\notebook;
 
 use app\libraries\Core;
-use app\models\gradeable\AbstractTextBox;
+use app\models\notebook\AbstractTextBox;
 
 /**
  * Class SubmissionTextBox
- * @package app\models\gradeable
+ * @package app\models\notebook
  */
 class SubmissionTextBox extends AbstractTextBox {
     public function __construct(Core $core, array $details) {
